smart_data_ingestor/requirements.txt
flask
requests
pandas
kafka-python
aiohttp